旧DB変換の金額をlong読みへ統一し関連伝票をcv10のIdへ張り替える
再変換に備えて変換ロジックを整備した。

getDataInt -> getDataLong (55箇所)
long のプロパティへ代入している箇所だけを置き換えた。明細金額合計/
上代合計/下代合計/内税消費税/外税消費税/値引1-3/明細の金額/消費税CD、
および long 側5クラスの関連伝票NO。getHeaderNebiki は戻り値が int の
ままだと Nebiki00Total(long) へ渡る前に溢れるため戻り値型も long にした。
RelateNo1 が int の Tran01Tenuri/Tran12Jyuchu/Tran13Hachu は getDataInt
のまま残す。旧データの実測は全て int 範囲内(ヘッダ合計 最大1.94億)だが、
旧列は NUMBER(10,0)/(14,0) で int を超えうるためプロパティ型に揃える。

関連伝票の張替 CnvTranRelateFix を追加
変換直後の RelateNo1 は旧SEQ_NOだが、アプリは RelateNo1 を cv10 の Id
として扱う(WHERE a.RelateNo1 = h.Id。発注残完了設定・入荷予定表・配分)。
張り替えないと変換データがこれらの機能で紐付かない。
変換順序は 売上→…→仕入→…→受注→発注 で依存(発注→仕入、受注→売上)と
逆のため各変換の中では親のIdを引けない。よって全変換の後に走る独立
ステップとして一括UPDATEし、親が未変換なら0件で終わる作りにした。
対象は 仕入→発注 / 売上→受注 / 移動受→積送。旧DBの関連伝票NOは仕入
8,602件中3件・売上50,314件中7件しか非0だが、移動受は170,930件全件が
積送を指しており実在率99.3%で、ここが実質最大の未張替だった。
判定は親側の OldSeqNo の一致のみで、値域が重ならない(Id は小さく
OldSeqNo は500万以上)ため再実行は0件で冪等。最小構成のDBで1対多・
親なし・RelateNo1=0・2回目0件を確認した。

Tran00Uriage.RelateNo1 / Tran03Shiire.RelateNo1 を int -> long
Tran05Ido/Tran10IdoOut/Tran11IdoIn は既に long で不統一だった。波及して
WarnIfLinkedZanCompletedAsync の引数と ZanCompletionRow.RelateNo1 も long
へ。どちらも発注/受注Idを保持する用途。

OldSeqNo に非ユニークキー nk7 を13テーブルへ追加
再変換と上記の張替で旧SEQ_NOから逆引きするため。nk1-nk6 は使用済み。

OldTableCommentAttr の誤りを修正
- TranKinHeader.KakeDay: 在庫計上日 -> 掛計上日 (2026-08-16の改名時の更新漏れ)
- Tran01Tenuri.Id_Customer: 顧客TELから解決したIdである旨を2引数形で明示
- Tran99MaterialMeisai.OldSeqNo: 明細への追加ミスのためプロパティを削除。
  OldSeqNo はヘッダ側3クラスへ整理。変換でも設定しておらず、JSON明細だが
  MissingMemberHandling 既定=Ignore かつ DefaultValueHandling.Ignore で
  0 は出力されないため既存データの読み込みに影響しない。
照合はプロパティ付与78件が対象で、71件は実在一致、3件は説明式のため対象外。
まともなスキーマ(^CV[0-9]{2,3}$ の77件、接続先CV163を含む)で確認し、
どのスキーマにも存在しない列名は0件だった。

入金/支払のKakeDayが在庫計上日を読んでいたのを修正
KakeDay は売掛・買掛の集計/元帳/消込の基準日で、売上・仕入は掛計上日を
読んでいる。区分6/7 は両列が全件同値(1件+2,299件)のため実害は出ていない。

+++ CvWpfclient/Helpers/ViewModels/BaseTranInputViewModel.cs
@@ -176,7 +176,7 @@ public abstract partial class BaseTranInputViewModel<TDen> : BasePlainLightMente
 	/// <param name="denLabel">保存した伝票の名称（"仕入" / "出荷"）</param>
 	/// <param name="zanLabel">紐付く残伝票の名称（"発注" / "受注"）</param>
 	/// <param name="settingLabel">案内する残完了設定画面の名称</param>
-	protected async Task WarnIfLinkedZanCompletedAsync(Type zanType, int relateNo1, string denLabel, string zanLabel, string settingLabel) {
+	protected async Task WarnIfLinkedZanCompletedAsync(Type zanType, long relateNo1, string denLabel, string zanLabel, string settingLabel) {
 		if (relateNo1 <= 0) {
 			return;
 		}
